backend/requirements.txt
psycopg2-binary==2.8.4
asgiref==3.2.3
Django==3.0.1
djangorestframework==3.11.0
pytz==2019.3
sqlparse==0.3.0
django-cors-headers==3.2.0
pillow==6.2.1
djangorestframework-camel-case==1.1.2
django-multiselectfield==0.1.11
